Разделение классов
